requirements.txt
streamlit
request
opencv-python
pillow
numpy
matplotlib